+++ package/Samples/Editor/SampleSettingsEditor.cs
@@ -12,17 +12,17 @@ namespace Samples
     [CustomEditor(typeof(SampleSettings))]
     public class SampleSettingsEditor : Editor
     {
-        static string kRequiresProperty = "m_RequiresItem";
-        static string kRuntimeToggleProperty  = "m_RuntimeToggle";
+        static string k_RequiresProperty = "m_RequiresItem";
+        static string k_RuntimeToggleProperty  = "m_RuntimeToggle";
 
-        static GUIContent kShowBuildSettingsLabel = new GUIContent("Build Settings");
-        static GUIContent kRequiresLabel = new GUIContent("Item Requirement");
+        static GUIContent k_ShowBuildSettingsLabel = new GUIContent("Build Settings");
+        static GUIContent k_RequiresLabel = new GUIContent("Item Requirement");
 
-        static GUIContent kShowRuntimeSettingsLabel = new GUIContent("Runtime Settings");
-        static GUIContent kRuntimeToggleLabel = new GUIContent("Should I stay or should I go?");
+        static GUIContent k_ShowRuntimeSettingsLabel = new GUIContent("Runtime Settings");
+        static GUIContent k_RuntimeToggleLabel = new GUIContent("Should I stay or should I go?");
 
-        bool showBuildSettings = true;
-        bool showRuntimeSettings = true;
+        bool m_ShowBuildSettings = true;
+        bool m_ShowRuntimeSettings = true;
 
         SerializedProperty m_RequiesItemProperty;
         SerializedProperty m_RuntimeToggleProperty;
@@ -32,25 +32,25 @@ namespace Samples
             if (serializedObject == null || serializedObject.targetObject == null)
                 return;
 
-            if (m_RequiesItemProperty == null) m_RequiesItemProperty = serializedObject.FindProperty(kRequiresProperty);
-            if (m_RuntimeToggleProperty == null) m_RuntimeToggleProperty = serializedObject.FindProperty(kRuntimeToggleProperty);
+            if (m_RequiesItemProperty == null) m_RequiesItemProperty = serializedObject.FindProperty(k_RequiresProperty);
+            if (m_RuntimeToggleProperty == null) m_RuntimeToggleProperty = serializedObject.FindProperty(k_RuntimeToggleProperty);
 
             serializedObject.Update();
-            showBuildSettings = EditorGUILayout.Foldout(showBuildSettings, kShowBuildSettingsLabel);
-            if (showBuildSettings)
+            m_ShowBuildSettings = EditorGUILayout.Foldout(m_ShowBuildSettings, k_ShowBuildSettingsLabel);
+            if (m_ShowBuildSettings)
             {
                 EditorGUI.indentLevel++;
-                EditorGUILayout.PropertyField(m_RequiesItemProperty, kRequiresLabel);
+                EditorGUILayout.PropertyField(m_RequiesItemProperty, k_RequiresLabel);
                 EditorGUI.indentLevel--;
             }
 
             EditorGUILayout.Space();
 
-            showRuntimeSettings = EditorGUILayout.Foldout(showRuntimeSettings, kShowRuntimeSettingsLabel);
-            if (showRuntimeSettings)
+            m_ShowRuntimeSettings = EditorGUILayout.Foldout(m_ShowRuntimeSettings, k_ShowRuntimeSettingsLabel);
+            if (m_ShowRuntimeSettings)
             {
                 EditorGUI.indentLevel++;
-                EditorGUILayout.PropertyField(m_RuntimeToggleProperty, kRuntimeToggleLabel);
+                EditorGUILayout.PropertyField(m_RuntimeToggleProperty, k_RuntimeToggleLabel);
                 EditorGUI.indentLevel--;
             }