+++ package/Samples/SampleConstants.cs
@@ -5,11 +5,11 @@ namespace Samples
     /// <summary>
     /// Static constants
     /// </summary>
-	public struct SampleConstants
-	{
+    public static class SampleConstants
+    {
         /// <summary>
         /// Key we use to store and retrieve custom configuration settings from EditorBuildSettings
         /// </summary>
-		public const string kSettingsKey = "com.unity.xr.management.sample_settings";
-	}
+        public const string k_SettingsKey = "com.unity.xr.management.sample_settings";
+    }
 }